requirements.txt
streamlit
python-docx
pandas
openpyxl
altair
requests
